feat(sample): resolve the samples a movie fragment declares into extents
`movie_fragment::sample_extents` resolves a `moof` against its movie by
the three-step fallback of §8.8.7 and §8.8.8 into `SampleExtent`s,
anchoring data per §8.8.7.1 and decode times per §8.8.12. It settles
every track fragment first — defaults, `data_reference_index` off the
`stsd` entry, where the track ends — and moves the caller's
`TrackDecodeTimes` before handing the extents over, so an empty-duration
fragment moves its track as §8.8.7.1 has it; `advance` goes.

Refs: #70

+++ crates/isobmff-sample/src/sample.rs
@@ -122,7 +122,7 @@ pub struct SampleExtent {
     sample_composition_time_offset: i64,
     sample_flags: u32,
     sample_description_index: u32,
-    data_reference_index: u32,
+    data_reference_index: u16,
     extent: Range<u64>,
 }
 
@@ -140,7 +140,7 @@ impl SampleExtent {
         sample_composition_time_offset: i64,
         sample_flags: u32,
         sample_description_index: u32,
-        data_reference_index: u32,
+        data_reference_index: u16,
         extent: Range<u64>,
     ) -> Self {
         Self {
@@ -193,7 +193,7 @@ impl SampleExtent {
 
     /// Returns the `dref` entry naming the resource the bytes of the sample lie in
     #[must_use]
-    pub const fn data_reference_index(&self) -> u32 {
+    pub const fn data_reference_index(&self) -> u16 {
         self.data_reference_index
     }
